ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
fasttext-wheel==0.9.2
numpy>=1.24.0,<2.0.0
pydantic>=2.0.0
